+++ client/client.ts
@@ -53,7 +53,7 @@ import { PathPageNavigator, parseRefFromURI } from "./navigator.ts";
 import { EventHook } from "./plugos/hooks/event.ts";
 import { Space } from "./space.ts";
 import { evalStatement } from "./space_lua/eval.ts";
-import { parseExpressionString, parse as parseLua } from "./space_lua/parse.ts";
+import { parseExpressionString, parseBlock as parseLua } from "./space_lua/parse.ts";
 import type { LuaCollectionQuery } from "./space_lua/query_collection.ts";
 import {
   LuaEnv,