+++ plug-api/syscalls/lua.ts
@@ -1,8 +1,14 @@
 import { syscall } from "../syscall.ts";
 import type { LuaBlock, LuaExpression } from "../../client/space_lua/ast.ts";
+import type { PrintOptions } from "../../client/space_lua/pretty_print.ts";
 
+export function parseBlock(code: string): Promise<LuaBlock> {
+  return syscall("lua.parseBlock", code);
+}
+
+/** @deprecated use {@link parseBlock} instead */
 export function parse(code: string): Promise<LuaBlock> {
-  return syscall("lua.parse", code);
+  return syscall("lua.parseBlock", code);
 }
 
 export function parseExpression(expression: string): Promise<LuaExpression> {
@@ -11,4 +17,18 @@ export function parseExpression(expression: string): Promise<LuaExpression> {
 
 export function evalExpression(expression: string): Promise<any> {
   return syscall("lua.evalExpression", expression);
+}
+
+export function prettyPrintBlock(
+  block: LuaBlock,
+  opts?: PrintOptions,
+): Promise<string> {
+  return syscall("lua.prettyPrintBlock", block, opts);
+}
+
+export function prettyPrintExpression(
+  expression: LuaExpression,
+  opts?: PrintOptions,
+): Promise<string> {
+  return syscall("lua.prettyPrintExpression", expression, opts);
 }